+++ src/Kit/DataEntry/DatePicker/RangePicker.tsx
@@ -1,4 +1,4 @@
-import {faCalendar, faCircleXmark, faClock} from '@fortawesome/free-regular-svg-icons';
+import {faCalendar, faClock, faCircleXmark, faArrowRight} from '@fortawesome/free-solid-svg-icons';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 import {KitInputWrapper} from '@kit/DataEntry/InputWrapper';
 import {useKitTheme} from '@theme/useKitTheme';
@@ -76,6 +76,7 @@ const KitRangePicker = forwardRef<any, IKitRangePicker>(
                                 : false
                         }
                         className={clx}
+                        separator={<FontAwesomeIcon icon={faArrowRight} />}
                     />
                 </div>
             </KitInputWrapper>